+++ backend/mvorganizer/crud/series.py
@@ -2,6 +2,8 @@ from sqlite3 import IntegrityError
 from sqlalchemy.orm import Session
 from typing import Optional, List
 
+from ..exceptions import DuplicateEntryException, InvalidIDException
+
 from .. import schemas
 from .. import models
 from .. import utils
@@ -24,26 +26,29 @@ def get_all_series(db: Session) -> List[models.Series]:
 
 
 def add_series(db: Session, name: str) -> models.Series:
-    series = models.Series(name=name)
+    series = models.Series(name=name, sort_name=utils.generate_sort_name(name))
 
     try:
         db.add(series)
         db.commit()
         db.refresh(series)
     except Exception as e:
-        logger.error(f"SqlAlchemy exception {str(e)}. Doing rollback")
         db.rollback()
-        return
+        raise DuplicateEntryException(f"Series {name} already exists")
 
     return series
 
 
-def delete_series(db: Session, series: models.Series) -> models.Series:
+def delete_series(db: Session, series_id: int) -> models.Series:
+    series = get_series(db, series_id)
+    if series is None:
+        raise InvalidIDException(f"Series with id {series_id} doesn't exists")
+
     try:
         db.delete(series)
         db.commit()
     except Exception as e:
         logger.error(f"Couldn't delete series {series}. {e} Doing rollback")
         db.rollback()
-        return None
+        raise e
     return series
